refactor: 계산 결과 타입을 long 으로 변경

+++ src/main/java/calculator/Controller/CalculateController.java
@@ -25,7 +25,7 @@ public class CalculateController {
         return inputView.inputStrings();
     }
 
-    private int calculateSum(String input) {
+    private long calculateSum(String input) {
         return Calculator.splitSumCalculator(input);
     }
 }
